pyproject.toml
[tool.poetry]
name = "iscc"
version = "1.0.5"
build = "build.py"
description = "ISCC: Reference Implementation"
authors = ["Titusz Pan <tp@py7.de>"]
license="BSD 2-clause"
readme = "README.md"
homepage = "https://iscc.codes/"
repository = "https://github.com/iscc/iscc-specs"
keywords=["iscc", "identifier", "media", "content", "hash", "blockchain", "similarity"]
classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
]

[tool.poetry.dependencies]
python = "^3.6"
xxhash = "^1"
Pillow = "^6"
blake3 = "^0.1.5"
cython = { version = "^0.29", optional = true }

[tool.poetry.extras]
cython = ["cython"]

[tool.poetry.dev-dependencies]
pytest = "^5"
mkdocs = "^1"
mkdocs-material = "^4"
mkdocs-redirects = "^1.0.0"
mkdocs-minify-plugin = "^0"
mkdocs-git-revision-date-localized-plugin = "^0"
black = "^19.10b0"
codetiming = "^1.2.0"
humanize = "^2.5.0"

[build-system]
requires = ["poetry>=1.0", "wheel"]
build-backend = "poetry.masonry.api"
